Fixed some embarrassing bugs in SF Controller.
Also, I've given up on Sylvester Control. Trash-tier numerical problems
that ain't nobody got time to deal with. Who wants to deal with a level
set function that takes an error of 1e-3 and scales it to 1e30! What is
that even! Ugh.

The next commit should be for reorganization but got to write report so
going to leave reorganization of code for later. Sorry! :'(

Additions:
 - Spline Arc Length estimation.
 - Now working with the campus world map. Yay!

Modifications:
 - Don't know wtf i was thinking with the frame derivative construction,
   but it was so obviously wrong. No wonder I was diverging. Ultra
   sucky.